Add sample subscription for monitoring LSA protection events
Windows 8.1 & Server 2012 R2 introduced a new opt-in security feature,
Additional LSA Protection, which improves the security of the Local
Security Authority (LSA) by configuring the LSASS (LSA Subsystem Service)
as a protected process. This substantially locks down access to the
process from other userland processes, which is of particular use in
defending against credential theft attacks (e.g. mimikatz).

If enabled it's useful to collect some related events:
 - Process startup events confirming LSASS started as a protected process
 - Code integrity events alerting of blocked drivers due to LSA policy

The latter category also applies to audit events generated in the event a
driver would have been blocked if Additional LSA Protection was enabled in
the case Audit Mode is configured in preparation for enabling the feature.

More details can be found on TechNet:
https://technet.microsoft.com/en-au/library/dn408187.aspx